[Feat] - 같은 루트 질문에 대한 다른 사람 답변 조회 기능 구현 (#237)
* feat: 루트 질문 참조 답변 DTO 및 조회 기능 추가

- RootQuestionReferenceAnswer DTO 추가 (nickname, interview_id, answer_content, answer_rank)
- AnswerRepository에 findTopAnswersByRootQuestionAndRank 메서드 추가
- 동일한 루트 질문에 대한 다른 사용자 답변을 인터뷰 좋아요 수 기준으로 조회
- A 랭크 우선, 부족 시 B 랭크로 최대 3개 답변 반환



* feat: 자신의 인터뷰 결과에 참조 답변 포함 기능 구현

- InterviewResultResponse에 rootQuestionReferenceAnswers 필드 추가
- InterviewService.findMyInterviewResult에서 참조 답변 조회 로직 통합
- 다른 사용자 인터뷰 결과 조회 시에는 참조 답변을 null로 반환
- MAX_REFERENCE_ANSWERS_COUNT 상수로 최대 개수 관리 (3개)



* test: 참조 답변 기능 테스트 및 픽스처 타입 수정

- AnswerFixtureBuilder의 likeCount 타입을 Integer에서 Long으로 수정
- InterviewControllerTest에 참조 답변 테스트 데이터 추가 (김철수, 이영희, 박민수)
- 인터뷰 좋아요 수 기준 정렬 검증 (10L, 5L, 3L)
- RestDocs 문서화에 참조 답변 필드 설명 추가
- AnswerControllerTest의 likeCount 사용법 수정



* chore: 클로드 관련 세팅 추가

* test: 리포지토리, 서비스 테스트 코드 추가

---------

+++ api/src/test/java/com/samhap/kokomen/global/fixture/answer/AnswerFixtureBuilder.java
@@ -12,7 +12,7 @@ public class AnswerFixtureBuilder {
     private String content;
     private AnswerRank answerRank;
     private String feedback;
-    private Integer likeCount;
+    private Long likeCount;
 
     public static AnswerFixtureBuilder builder() {
         return new AnswerFixtureBuilder();
@@ -43,7 +43,7 @@ public class AnswerFixtureBuilder {
         return this;
     }
 
-    public AnswerFixtureBuilder likeCount(Integer likeCount) {
+    public AnswerFixtureBuilder likeCount(Long likeCount) {
         this.likeCount = likeCount;
         return this;
     }
